refactor: Convert tests & configs to import syntax
