+++ test/cases/common-async/webpack.config.js
@@ -1,5 +1,5 @@
-const webpack = require('webpack');
-const ExtractTextPlugin = require('../../../src/index');
+import webpack from 'webpack';
+import ExtractTextPlugin from '../../../src/index';
 
 module.exports = {
   entry: './index',
