+++ test/cases/multiple-entries-filename/webpack.config.js
@@ -1,4 +1,5 @@
-const ExtractTextPlugin = require('../../../src/');
+import ExtractTextPlugin from '../../../src/index';
+
 module.exports = {
   entry: {
     'js/a': './a',
