+++ test/schema-validation.test.js
@@ -1,8 +1,6 @@
-// import ExtractTextPlugin from '../src';
-// const loader = require.resolve('../src/loader');
-/* eslint-disable */
-var ExtractTextPlugin = require("../src/");
-var loader = require.resolve('../src/loader');
+/* eslint-disable no-unused-expressions */
+import ExtractTextPlugin from '../src';
+
 
 describe('json schema validation', () => {
   it('does not throw if a filename is specified', () => {
